Added r0' integrand plotting function.
Added a function to plot the integrand against r0' for a set of values
and over a grid of values. Funciton appears exponential, suggesting an
integration method that is open. An extended midpoint version of
Simpson's rule could work. Began coding up the c++ framework, but need
to re-think. Consider passing params struct by reference through
multiple levels of integration.